Completed exercise 3.16
I have added the errorHandler middleware for the CastError error after routing, but I am unsure wether I was supposed to only add handling for this specific error (like we did in the course app) or for other error types as well. I have not yet implemented the GET operation for the /api/persons/:id endpoint which is the only operation that might return a CastError so this middleware is not being used right now.